+++ src/torchtime/constants.py
@@ -72,8 +72,9 @@ PHYSIONET_2012_VARS: Final[list] = [
     "ICUType3",
     "ICUType4",
 ]
-PHYSIONET_2012_CATEGORICAL: Final[list] = [20, 39, 41, 42, 43, 44]
 PHYSIONET_2012_STATIC: Final[list] = [38, 39, 40, 41, 42, 43, 44]
+PHYSIONET_2012_CATEGORICAL: Final[list] = [20, 39, 41, 42, 43, 44]
+PHYSIONET_2012_ORDINAL: Final[list] = [11]
 PHYSIONET_2012_MEANS: Final[dict] = {20: 0.0, 41: 0.0, 42: 0.0, 43: 1.0, 44: 0.0}
 
 # PhysioNet 2019
@@ -81,6 +82,7 @@ PHYSIONET_2019_DATASETS: Final[dict] = {
     "training": "https://archive.physionet.org/users/shared/challenge-2019/training_setA.zip",  # noqa: E501
     "training_setB": "https://archive.physionet.org/users/shared/challenge-2019/training_setB.zip",  # noqa: E501
 }
-PHYSIONET_2019_CATEGORICAL: Final[list] = [35, 36, 37]
 PHYSIONET_2019_STATIC: Final[list] = [34, 35, 36, 37, 38, 39]
+PHYSIONET_2019_CATEGORICAL: Final[list] = [35, 36, 37]
+PHYSIONET_2019_ORDINAL: Final[list] = []
 PHYSIONET_2019_MEANS: Final[dict] = {}