+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * ``n_channels`` attribute - MORE?
 * Split out static channels
+* Ordinal variables imputed correctly
 
 ### Changed